src/NoWoL.TestUtils/ObjectCreators/ValueTypeCreator.cs
using System;
using System.Collections.Generic;

namespace NoWoL.TestingUtilities.ObjectCreators
{
    /// <summary>
    /// Provides a way to create value types
    /// </summary>
    public class ValueTypeCreator : IObjectCreator
    {
        internal const string DefaultStringValue = "SomeValue";

        /// <summary>
        /// Determines whether this instance can create the specified object type.
        /// </summary>
        /// <param name="type">Type of the object.</param>
        /// <returns><c>true</c> if this instance can create the specified object type; otherwise, <c>false</c>.</returns>
        public bool CanHandle(Type type)
        {
            if (type == null)
            {
                throw new ArgumentNullException(nameof(type));
            }

            return type.IsValueType || type == typeof(string);
        }

        /// <summary>
        /// Create an instance of the specified object type.
        /// </summary>
        /// <param name="type">Type of the object.</param>
        /// <param name="objectCreators">A list of <see cref="IObjectCreator"/> to handle creation of sub objects.</param>
        /// <returns>The created object.</returns>
        public object Create(Type type, ICollection<IObjectCreator> objectCreators)
        {
            if (type == null)
            {
                throw new ArgumentNullException(nameof(type));
            }

            if (type == typeof(string))
            {
                return DefaultStringValue;
            }

            if (type.IsValueType)
            {
                return type.GetDefault();
            }

            throw new UnsupportedTypeException("Expecting a string or value type however received " + type.FullName);
        }
    }
}